+++ AdjustablePortals/AdjustablePortals.cs
@@ -12,23 +12,25 @@ namespace AdjustablePortals
     [BepInPlugin(PluginGUID, PluginName, PluginVersion)]
     [BepInDependency(Jotunn.Main.ModGuid)]
     [NetworkCompatibility(CompatibilityLevel.EveryoneMustHaveMod, VersionStrictness.Minor)]
+    [BepInDependency("org.bepinex.plugins.targetportal", BepInDependency.DependencyFlags.SoftDependency)]
     internal class AdjustablePortals : BaseUnityPlugin
     {
         public const string PluginGUID = "MidnightsFX.AdjustablePortals";
         public const string PluginName = "AdjustablePortals";
-        public const string PluginVersion = "0.1.0";
+        public const string PluginVersion = "0.3.0";
         internal static Harmony Harmony = new Harmony(PluginGUID);
 
         public static CustomLocalization Localization = LocalizationManager.Instance.GetLocalization();
         internal static ManualLogSource Log;
 
-        public void Awake()
-        {
+        public void Awake() {
             Log = this.Logger;
             new ValConfig(Config);
             TeleportItems.SetupTeleportLists();
+            Compatibility.CheckModCompat();
             Assembly assembly = Assembly.GetExecutingAssembly();
             Harmony.PatchAll(assembly);
+            Compatibility.TargetPortalCompat();
         }
     }
 }